refactor: 테스트 코드 리팩토링
코드 정렬 및 test 코드 리팩토링
